+++ R/init.R
@@ -159,7 +159,7 @@ init <- function(data,
     ui = ui_teal_with_splash(id = id, data = data, title = title, header = header, footer = footer),
     server = function(input, output, session) {
       # copy object so that load won't be shared between the session
-      data <- data$clone(deep = TRUE)
+      data <- data$copy(deep = TRUE)
       srv_teal_with_splash(
         input, output, session,
         data = data, modules = modules, filter = filter